refactor optimizers into two files
